requirements.txt
Flask==2.3.2
Pillow==10.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
